loader/requirements.txt
influx-client
requests
paho-mqtt